refactor(desktop): split app shell daily review bridge
