+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -9,6 +9,7 @@ const sourcePaths = [
   'app.tsx',
   'app-shell-quick-chat-actions.ts',
   'app-shell-layout-actions.ts',
+  'app-shell-daily-review-bridge.ts',
   'app-shell-daily-review-actions.ts',
   'app-shell-turn-actions.ts',
   'app-shell-turn-view-model.ts',
